requirements.txt
django>=3.2.18
psycopg2-binary>=2.9
djangorestframework>=3.14
markdown>=3.4
django-filter>=22.1
# djangorestframework-jwt
# psycopg2>=2.9